refactor : default path url
