--- src/app/flyover/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Flyover() {
-  return (
-    <div>Flyover</div>
-  )
-}
-
-export default Flyover
