+++ src/app/invoices/form.tsx
@@ -1,11 +1,11 @@
 "use client";
 import Input from "@/components/Input";
-import React, { useState } from "react";
+import React, { useEffect, useState, Fragment } from "react";
 import InvoiceItem from "./invoice-items";
-import useItem from "@/hooks/useItem";
-import { formatNumber } from "@/utils";
+import { formatNumber, generateInvoice } from "@/utils";
 import { Details, Item } from "@/@types/types";
-import useDetail from "@/hooks/useDetail";
+import { invDetail } from "@/utils/detailInvoice";
+import useDimensions from "@/hooks/useDimension";
 
 type IProps = {
   Items: Item[];
@@ -27,8 +27,8 @@ function Form({
   Details,
   setDetails,
 }: IProps) {
+  const { height, width } = useDimensions();
   const formatTotal = formatNumber(totalPrice());
-  const formatFinalTotal = formatNumber(totalPrice() - Details.discount);
 
   const onChangeDetails = (
     event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
@@ -37,9 +37,10 @@ function Form({
   };
 
   return (
-    <div className="w-full">
+    <div className="w-full text-sm">
       <div className="flex flex-row">
         <Input
+          disable={true}
           onChange={(e) => onChangeDetails(e)}
           value={Details?.invoice_number}
           placeholder="#0001"
@@ -48,96 +49,71 @@ function Form({
           type="text"
         />
       </div>
-      <div className="flex flex-row mt-4">
-        <Input
-          onChange={(e) => onChangeDetails(e)}
-          value={Details?.bill_from}
-          placeholder="Company Name, depok street no 77."
-          title="Your company details"
-          name="bill_from"
-          type="text-area"
-        />
-        <div className="mx-4" />
-        <Input
-          onChange={(e) => onChangeDetails(e)}
-          value={Details?.bill_to}
-          placeholder="Company Name, ciracas street no 145."
-          title="Bill to"
-          name="bill_to"
-          type="text-area"
-        />
+      <div className="mt-4 flex flex-col items-center lg:flex-row">
+        <div className="w-full lg:w-1/2">
+          <div className="text-sm font-semibold">{invDetail.storeName}</div>
+          <div>{invDetail.storeAddress}</div>
+          <div>
+            {invDetail.phoneNumber} - {invDetail.phoneNumber2}
+          </div>
+        </div>
+        <div className="mt-4 w-full lg:mt-0 lg:w-1/2">
+          <Input
+            onChange={(e) => onChangeDetails(e)}
+            value={Details?.bill_to}
+            placeholder="Nama Pelanggan"
+            title="Nama Pelanggan"
+            name="bill_to"
+            type="text"
+          />
+          <div className="my-2"></div>
+          <Input
+            onChange={(e) => onChangeDetails(e)}
+            title="Tanggal"
+            type="date"
+            name="date"
+          />
+        </div>
       </div>
-      <div className="flex flex-row mt-4">
-        <Input
-          onChange={(e) => onChangeDetails(e)}
-          title="Date issued"
-          type="date"
-          name="date"
-        />
-        <div className="mx-4" />
-        <Input
-          onChange={(e) => onChangeDetails(e)}
-          title="Due date"
-          type="date"
-          name="due_date"
-        />
-      </div>
+
       {/* items */}
-      <div className="bg-gray-100 mt-8 rounded p-4">
+      <div className="mt-8 rounded bg-gray-100 p-4">
         {/* form grup */}
         {Items.map((item, index) => {
+          const currentWidth = width;
+          const showHeader =
+            currentWidth < 1024 ? true : index == 0 ? true : false;
           return (
             <InvoiceItem
               key={item.id}
               item={item}
-              showHeader={index == 0 ? true : false}
+              showHeader={showHeader}
               onChangeItem={setItems}
               onRemoveItem={() => removeItem(item.id)}
             />
           );
         })}
-        <div className="w-full flex justify-center my-4">
+        <div className="my-4 flex w-full justify-center">
           <button
             onClick={() => addNewItem()}
-            className="font-semibold text-sm text-white bg-blue-600 rounded px-4 py-2"
+            className="rounded bg-blue-600 px-4 py-2 text-sm font-semibold text-white"
           >
             Add Item
           </button>
         </div>
       </div>
       {/* end items */}
-      <div className="flex flex-row items-start mt-12">
-        <div className="w-1/2">
-          <Input
-            onChange={(e) => onChangeDetails(e)}
-            value={Details?.notes}
-            placeholder="Notes"
-            title="Notes"
-            type="text-area"
-            name="notes"
-          />
-        </div>
-        <div className="w-1/2 flex flex-row justify-end items-center text-sm">
-          <div className="pr-4 text-gray-500 font-semibold">
+      <div className="mt-12 flex flex-row items-start">
+        <div className="w-1/2"></div>
+        <div className="flex w-1/2 flex-row items-center justify-end text-sm">
+          <div className="pr-4 font-semibold text-gray-500">
             <div className="pb-2">Sub Total</div>
-            <div className="pb-2">Discount</div>
-            <div className="pt-4 text-blue-600">Total</div>
+            {/* <div className="pb-2">Discount</div> */}
+            <div className="pt-3 text-blue-600">Total</div>
           </div>
-          <div className="text-end items-end">
-            <div className="font-bold text-lg pb-2">{formatTotal}</div>
-            <div>
-              <Input
-                onChange={(e) => onChangeDetails(e)}
-                value={Details?.discount.toString()}
-                placeholder="Discount"
-                type="number"
-                name="discount"
-                styles="text-sm text-end"
-              />
-            </div>
-            <div className="pt-3 text-lg text-blue-600">
-              {formatFinalTotal ? formatFinalTotal : 0}
-            </div>
+          <div className="items-end text-end">
+            <div className="pb-2 text-lg font-bold">{formatTotal}</div>
+            <div className="pt-2 text-lg text-blue-600">{formatTotal}</div>
           </div>
         </div>
       </div>
